+++ aim/ql/tree/__init__.py
@@ -1,0 +1,2 @@
+from aim.ql.tree.abstract_syntax_tree import AbstractSyntaxTree
+from aim.ql.tree.binary_expression_tree import BinaryExpressionTree